jwt authentication

+++ controllers/blogs.js
@@ -1,8 +1,18 @@
 const blogsRouter = require("express").Router();
+const jwt = require("jsonwebtoken");
 const Blog = require("../models/blog");
+const User = require("../models/user");
 
-blogsRouter.get("/", async (request, response, next) => {
-  const blogs = await Blog.find({});
+const getTokenFrom = request => {
+  const authorization = request.get("authorization");
+  if (authorization && authorization.toLowerCase().startsWith("bearer ")) {
+    return authorization.substring(7);
+  }
+  return null;
+};
+
+blogsRouter.get("/", async (_, response, next) => {
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   return response.json(blogs);
 });
 
@@ -16,14 +26,16 @@ blogsRouter.get("/:id", async (request, response, next) => {
 });
 
 blogsRouter.put("/:id", async (request, response, next) => {
-  if (!request.body.title || !request.body.url) {
+  const body = request.body;
+  if (!body.title || !body.url) {
     return response.send(400).json({ error: "name or url missing" });
   }
   const blog = {
-    title: request.body.title,
-    author: request.body.author,
-    url: request.body.url,
-    likes: request.body.likes || 0
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes || 0,
+    date: new Date()
   };
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
     new: true
@@ -32,18 +44,44 @@ blogsRouter.put("/:id", async (request, response, next) => {
 });
 
 blogsRouter.post("/", async (request, response, next) => {
-  if (!request.body.title && !request.body.url) {
-    return response.status(400).json({ error: "name or url missing" });
+  // if (!request.body.title && !request.body.url) {
+  //   return response.status(400).json({ error: "name or url missing" });
+  // }
+  const body = request.body;
+
+  const token = getTokenFrom(request);
+
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
   }
-  if (!request.body.likes) request.body.likes = 0;
 
-  const blog = new Blog(request.body);
+  const user = await User.findById(decodedToken.id);
 
-  const result = await blog.save();
-  return response.status(201).json(result);
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes || 0,
+    date: new Date(),
+    user: user._id
+  });
+
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.json(savedBlog);
 });
 
 blogsRouter.delete("/:id", async (request, response, next) => {
+  const token = getTokenFrom(request);
+
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   return response.status(204).end();
 });
